refactor: type -> interface 로 변경
